refactor(#4794): extract session transition locking
Keep Discord boot wiring below its frozen production LoC while preserving transition behavior and tests verbatim.
