+++ src/services/discord/router/message_handler/intake_turn/race_loss/requeue_tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 
 struct EnvReset(Option<std::ffi::OsString>);
